Refactor to use create_app factory
