+++ tests/test_download_payslip.py
@@ -6,17 +6,19 @@ import pytest
 os.environ['DATABASE_URL'] = 'sqlite:///:memory:'
 os.environ['SESSION_SECRET'] = 'testing'
 
-import app
-from app import db
+from app import create_app, db
 from models import Role, User, Employee, PayPeriod, Payroll, PayrollEntry
+
+
+app = create_app()
 
 
 @pytest.fixture
 def client():
-    app.app.config['TESTING'] = True
-    app.app.config['SQLALCHEMY_DATABASE_URI'] = 'sqlite:///:memory:'
+    app.config['TESTING'] = True
+    app.config['SQLALCHEMY_DATABASE_URI'] = 'sqlite:///:memory:'
 
-    with app.app.app_context():
+    with app.app_context():
         db.drop_all()
         db.create_all()
 
@@ -69,10 +71,10 @@ def client():
         db.session.add(entry)
         db.session.commit()
 
-    with app.app.test_client() as client:
+    with app.test_client() as client:
         client.post('/login', data={'username': 'admin', 'password': 'password'})
         yield client, payroll.id
-        with app.app.app_context():
+        with app.app_context():
             db.session.remove()
             db.drop_all()
